+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.1.3
+* [DEL] Deprecate delays when sending commands and set their default value to 0
+* [IMP] Support custom icons
+* [DEL] Deprecate the `setFont` function and send font size and color through Bluetooth
+* [ADD] Add SVG icons and scaling factor
+* [FIX] Replace the old rename PNG icon with the new one
+* [IMP] Improve code documentation
+* [IMP] Add support for wrapping TextDrawings on multiple lines
+
 ## 1.1.2
 * [ADD] Add size and color parameters to TextDrawing
 * [ADD] Convert WhiteScreen into BlankScreen with a color parameter